chores: changed position type from float to int

+++ inc/Platform.h
@@ -11,8 +11,8 @@ extern bool is_debug;
 class Platform{
 public:
     struct CollisionShape {
-        float y;
-        float left, right;
+        int y;
+        int left, right;
     };
 
 public:
