+++ src/Platform.cpp
@@ -7,6 +7,6 @@ void Platform::OnDraw(const Camera& camera) const {
         setfillcolor(RGB(0, 0, 255));
         solidcircle(render_position.x, render_position.y, 3);
         setlinecolor(RGB(255, 0, 0));
-        Line(camera, (int)shape.left, (int)shape.y, (int)shape.right, (int)shape.y);
+        Line(camera, shape.left, shape.y, shape.right, shape.y);
     }
 }
